+++ packages/brick_offline_first_with_rest/lib/src/offline_first_with_rest_repository.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:brick_offline_first_with_rest/src/models/offline_first_with_rest_model.dart';
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_queue_client.dart';
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_request_queue.dart';